Get rid of a lot of the bufsize noise
Rather than passing the bufsize around as an argument to every single
function, we now encapsulate it in a `PipelineConfig` struct. This has
the upside of avoiding almost all of the visual noise caused by having
ot pass it around all of the time, at the expense of actually
customising its value be much more painful. For instance, the current
API used like this:

    Pipeline::new(whatever).config(PipelineConfig::default().buf_size(10))
        .map(something_else)

actually changes the size of the buffer between `whatever` and
`something` else, *not* the buffer to the initial seeding iterator
`whatever`. In fact, a side-effect of this change is that there is now
*no way* to change the size of the buffer that `whatever` feeds.
Similarly, `Multiplexer` can now only use the default configuration and
is unable to customise the size of its internal buffers at all.

While I'm much happier with having a configuration method like this in
general I'm not especially happy with how awkward it is to use currently
and will keep playing with some other shapes of it