Send messages between stages in batches
Sending messages in batches reduces the amount of time spent in mpsc lock
contention since N-1 of the time you already have the message available rather
than having to go out to the queue for it. But it can reduce effective use of
multiple cores because a worker has to wait for all N messages to become
available rather than being able to start work on the N-1 it would already have
already had. To be able to tune for both of these cases, the batch size is
configureable via the `config` method on the Pipeline

Both of these runs are on the benchmark suite that's modified as of this
commit. To get it to run on the previous commit you'll need to comment
out the `.config` lines but it will otherwise run unmodified

Before:

    test bench_heavy_multi  ... bench:  12,195,560 ns/iter (+/- 733,274)
    test bench_heavy_single ... bench:  36,972,715 ns/iter (+/- 3,515,289)
    test bench_light_multi  ... bench: 462,368,875 ns/iter (+/- 118,079,961)
    test bench_light_single ... bench: 209,381,097 ns/iter (+/- 15,562,339)

After:

    test bench_heavy_multi  ... bench:  13,272,238 ns/iter (+/- 548,308)
    test bench_heavy_single ... bench:  49,700,229 ns/iter (+/- 4,172,453)
    test bench_light_multi  ... bench:  48,929,034 ns/iter (+/- 4,925,751)
    test bench_light_single ... bench:  37,529,783 ns/iter (+/- 4,777,694)

As you can see, the case of a small number of heavy messages is
marginally worse, but the case of lots of small messages is way better